requirements.txt
streamlit>=1.32.0
openai>=1.14.0
pandas>=2.2.1
plotly>=5.20.0
faker>=24.2.0
python-dotenv>=1.0.1
